pyproject.toml
[build-system]
requires = ["setuptools>=61", "setuptools_scm[toml]>=7"]
build-backend = "setuptools.build_meta"


[project]
name = "kwneuro"
authors = [
  { name = "Ebrahim Ebrahim", email = "ebrahim.ebrahim@kitware.com" },
  { name = "Sadhana Ravikumar", email = "sadhana.ravikumar@kitware.com" },
  { name = "David Allemang", email = "david.allemang@kitware.com" }
]
description = "Processing pipelines to extract brain microstructure from diffusion MRI"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "click >=8.1",
  "numpy <2.4",
  "dipy >=1.9",
  "pandas",
  "antspyx>=0.6.2",
]

[project.optional-dependencies]
hdbet = [
  "hd-bet == 2.0.1",
]
tractseg = [
  "TractSeg",
]
noddi = [
  "dmri-amico == 2.1.1",
  # Required because older amico/setuptools dependency chain
  # imports backports.tarfile (can be removed if amico drops it)
  "backports.tarfile",
]
combat = [
  "neuroCombat ==0.2.12",
]
antspynet = [
  "antspynet",
]
all = [
  "kwneuro[hdbet,tractseg,noddi,combat,antspynet]",
]
test = [
  "pytest >=6",
  "pytest-cov >=3",
  "pytest-mock >=3.14",
  "kwneuro[combat]",
]
dev = [
  "kwneuro[test,all]",
  "pre-commit",
  "pylint",
]
notebooks = [
  "matplotlib",
  "jupytext",
  "ipykernel",
  "notebook",
  "ipywidgets",
  "openneuro-py",
]
docs = [
  "sphinx>=7.0",
  "sphinx-autobuild==2024.10.3",
  "sphinx-autoapi>=3.0",
  "myst_parser>=0.13",
  "sphinx_copybutton",
  "sphinx_click",
  "furo>=2023.08.17",
]

[project.urls]
Homepage = "https://github.com/brain-microstructure-exploration-tools/kwneuro"
"Bug Tracker" = "https://github.com/brain-microstructure-exploration-tools/kwneuro/issues"
Discussions = "https://github.com/brain-microstructure-exploration-tools/kwneuro/discussions"
Changelog = "https://github.com/brain-microstructure-exploration-tools/kwneuro/releases"

[project.scripts]
gen_masks = "kwneuro.run:gen_masks"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/kwneuro/_version.py"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
  "ignore:.*legacy descoteaux07 SH basis.*:PendingDeprecationWarning",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]


[tool.coverage]
run.source = ["kwneuro"]
report.exclude_also = [
  '\.\.\.',
  'if typing.TYPE_CHECKING:',
]

[tool.mypy]
files = ["src", "tests"]
python_version = "3.10"
warn_unused_configs = true
strict = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "kwneuro.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["neuroCombat.*", "HD_BET.*", "tractseg.*", "amico.*", "antspynet.*"]
ignore_missing_imports = true


[tool.ruff]
src = ["src"]

[tool.ruff.lint]
extend-select = [
  "B",        # flake8-bugbear
  "I",        # isort
  "ARG",      # flake8-unused-arguments
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "ICN",      # flake8-import-conventions
  "G",        # flake8-logging-format
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
  "EXE",      # flake8-executable
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
  "PLC3002",  # Direct call of lambda -- this can often aid code clarity in my opinion
  "PLC0415",  # import-outside-toplevel -- lazy imports are an intentional pattern for optional deps
]
isort.required-imports = ["from __future__ import annotations"]
# Uncomment if using a _compat.typing backport
# typing-modules = ["kwneuro._compat.typing"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"scripts/**" = ["T20"]


[tool.pylint]
py-version = "3.10"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "import-outside-toplevel",
  "line-too-long",
  "missing-module-docstring",
  "wrong-import-position",
]
